refactor: split main.rs into modules and extract tests
Why:
As the codebase grew, main.rs became bloated (over 1,500 lines) with
mixed responsibilities: LSP handlers, Zsh process management,
document syncing, and integration test harnesses. This hindered
maintainability and made adding new LSP features difficult.

What:
- Create src/lib.rs and expose Backend and submodules
- Move document position/sync logic to src/document.rs
- Move Zsh capture daemon and completion parsing to src/completion.rs
- Move LanguageServer trait implementation to src/server.rs
- Simplify src/main.rs to CLI entry point only
- Extract TestClient and test harnesses into tests/common/mod.rs
- Separate integration tests into tests/server_test.rs and
  tests/completion_test.rs
- Update AGENTS.md and docs/ARCHITECTURE.md project structure docs
