+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod completion;
+pub mod document;
+pub mod server;
+
+pub use server::Backend;
